+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # OMC3 Changelog
 
+#### 2022-02-23
+
+- Fixed:
+  - The coupling calculation now includes additional columns in the output dataframes, which were missing while being needed later on by the correction calculation.
+
 #### 2022-01-10
 
 - Fixed:
   - Sequences for K-Modulation are now included in PyPi package
   - Bug fixed where default errors in K-Modulation would not have been taken into account
-  
+
 #### 2021-07-14 
 _by jdilly_